+++ packages/react/src/spatialized-container/SpatializedStatic3DElementContainer.tsx
@@ -83,12 +83,21 @@ function collectSources(children: React.ReactNode): ModelSource[] {
 }
 
 function SpatializedContent(props: SpatializedStatic3DContentProps) {
-  const { src, children, spatializedElement, onLoad, onError, autoPlay, loop } =
-    props
+  const {
+    src,
+    poster,
+    children,
+    spatializedElement,
+    onLoad,
+    onError,
+    autoPlay,
+    loop,
+  } = props
 
   const portalInstanceObject = useContext(PortalInstanceContext)!
 
   const modelURL = useMemo(() => getAbsoluteURL(src), [src])
+  const posterURL = useMemo(() => getAbsoluteURL(poster), [poster])
   const sources = useMemo(() => collectSources(children), [children])
 
   useEffect(() => {
@@ -100,8 +109,9 @@ function SpatializedContent(props: SpatializedStatic3DContentProps) {
       sources,
       autoplay: autoPlay,
       loop,
+      posterURL: posterURL ?? '',
     })
-  }, [modelURL, JSON.stringify(sources), autoPlay, loop])
+  }, [modelURL, JSON.stringify(sources), autoPlay, loop, posterURL])
 
   useEffect(() => {
     if (onLoad) {
@@ -150,7 +160,6 @@ function SpatializedStatic3DElementContainerBase(
   const extraRefProps = useCallback(
     (domProxy: SpatializedStatic3DElementRef) => {
       let modelTransform = new DOMMatrixReadOnly()
-
       return {
         get currentSrc(): string {
           const spatializedElement = (domProxy as any).__spatializedElement as
@@ -193,6 +202,26 @@ function SpatializedStatic3DElementContainerBase(
             | SpatializedStatic3DElement
             | undefined
           return spatializedElement?.paused ?? true
+        },
+        get duration(): number {
+          const spatializedElement = (domProxy as any).__spatializedElement as
+            | SpatializedStatic3DElement
+            | undefined
+          return spatializedElement?.duration ?? 0
+        },
+        get playbackRate(): number {
+          const spatializedElement = (domProxy as any).__spatializedElement as
+            | SpatializedStatic3DElement
+            | undefined
+          return spatializedElement?.playbackRate ?? 1
+        },
+        set playbackRate(value: number) {
+          const spatializedElement = (domProxy as any).__spatializedElement as
+            | SpatializedStatic3DElement
+            | undefined
+          if (spatializedElement) {
+            spatializedElement.playbackRate = value
+          }
         },
       }
     },