+++ packages/react/src/spatialized-container/types.ts
@@ -101,6 +101,7 @@ export type SpatializedStatic3DContainerProps =
   SpatialEventProps<SpatializedStatic3DElementRef> &
     Omit<React.ComponentPropsWithoutRef<'div'>, 'onLoad' | 'onError'> & {
       src?: string
+      poster?: string
       autoPlay?: boolean
       loop?: boolean
       children?: React.ReactNode
@@ -112,6 +113,7 @@ export type SpatializedStatic3DContainerProps =
 export type SpatializedStatic3DContentProps = {
   spatializedElement: SpatializedStatic3DElement
   src?: string
+  poster?: string
   autoPlay?: boolean
   loop?: boolean
   children?: React.ReactNode
@@ -142,6 +144,8 @@ export type SpatializedStatic3DElementRef = SpatializedDivElementRef & {
   play(): Promise<void>
   pause(): Promise<void>
   readonly paused: boolean
+  readonly duration: number
+  playbackRate: number
 }
 
 type CurrentTarget<T extends SpatializedElementRef> = {