+++ packages/visionOS/web-spatial/JSBCommand.swift
@@ -266,6 +266,8 @@ struct UpdateSpatializedStatic3DElementProperties: SpatializedElementProperties 
     let autoplay: Bool?
     let loop: Bool?
     let animationPaused: Bool?
+    let playbackRate: Double?
+    let posterURL: String?
 }
 
 struct UpdateSpatializedDynamic3DElementProperties: SpatializedElementProperties {